sql代码自动生成

+++ src/main/java/com/example/demo1/controller/PublishController.java
@@ -43,7 +43,7 @@ public class PublishController {
     public  String doPublish(@RequestParam(name = "title", required = false) String title,
                              @RequestParam(name = "description", required = false) String description,
                              @RequestParam(name = "tag" , required = false) String tag,
-                             @RequestParam(name = "id" ,required = false) Long id,
+                             @RequestParam(name = "id" ,required = false) Integer id,
                              HttpServletRequest request,
                              Model model){
 
@@ -84,7 +84,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String getPublish(@PathVariable(name = "id") Long id,
+    public String getPublish(@PathVariable(name = "id") Integer id,
                              Model model){
 
         QuestionDTO questionDTO = questionService.getById(id);
